MultipleCurves: Change weight formula
From now on, the angles at the segment's endpoints and the segment's
length will be used to calculate the weight.

Also segments can be replaced during post-processing. Consequently,
the order of iterating through the list is almost not relevant any
more; the algorithm ensures that the output looks good regardless
of how they are ordered.