MultiCurves: Update some comments
See my notes on paper for more details;
Most of the test cases look good. Most of the partially failing test
cases are caused by the following major TODOs (in order of priority):

- angle
- maxlen
- one
- magic

"angle" must check whether "fixing" the garph by replacing segments
actually increases the quality, based on the angle as a quantitative
measurement.

"maxlen" must be applied at all stages. It puts a limit on the maximum
length of a segment. Preferably it accounts for local changes, such
as the average segment length in the current shape.
Statistics may help with implementing the latter part correctly.

"one" is just checking all nodes with one endpoint. An idea is to
first find all of these nodes, then try to connect them if possible
(with custom replace decision logic if needed).

"magic" is needed to solve some difficult problems. E.g. the
multiCloseCurves test case fails because of the proximity of the nodes
at the center. This could be solved by accounting for the direction of
a path. Similar for multiEnclosedShapes.